research-service/pyproject.toml
[project]
name = "research-service"
version = "0.1.0"
description = "Mistral function-calling research agent for sports prediction markets"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.134.0",
    "httpx>=0.28.1",
    "mistralai>=1.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
